app/requirements.txt
async-generator==1.10
attrs==21.4.0
backports.entry-points-selectable==1.1.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
distlib==0.3.3
filelock==3.3.2
h11==0.12.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
numpy==1.21.4
outcome==1.1.0
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==1.2.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.10.0
wsproto==1.0.0
